Fix storage layer on Linux
Using O_APPEND and pwrite(2) doesn't mix well on Linux. It has a
(documented) bug that basically says that Linux ignores the offset
provided in pwrite(2) and just appends the bytes at the end of the file.
This broke Index::overwrite_date. To resolve this we're force to use
O_RDWR instead.