Change log target of switcher used actors returned errors
When running the coordinator server or participant dispatcher the
switcher would log any errors it encountered, but using a log target
that pointed to peers::switcher, which is misleading. These helper
run_actor function ensure that the log target is set to the correct
module. Note that could also be done by manually setting the target when
logging, but that doesn't set the module.