Add initial usage of request passport to HTTP module
Parses the "X-Request-ID" header from the request and uses it in the
passport if its valid. Adds the request-id header to the response and
uses it in logging the request.

Also adds a new constant N_RESPONSE_HEADERS to the cli which contains
the number of response headers to expect. Rather then having that
hard-coded in each binary individually.