Fix full synchronisation
Before it would incorrectly keep track of the number of locally missing
blobs, it would simply add it to a vector. This causes keys to be
included in it multiple times, causing the same blob to retrieved
multiple times.

A second fix is that it now checks if a timestamp for a retrieve blob is
removed or invalid, in which case it check if the blob length is zero.
Before it would run with the same assumption, but it would never check
it. This causes some fun errors where the blob length was huge (as it
would read an invalid length), causing reading errors later own.

These two bugs are fixed and some tests are added.