Make blob_already_stored_after_initial_check less flaky
Because the two requests were send right after each other it was
possible for the 2PC query B to retrieve a blob first, causing the test
to fail (as it expect 2PC query A to retrieve the blob first). Adding a
small timeout between the two requests should fix the order.