Fix incorrect file offset in UncommittedBlob
In Data::add_blob it would use the `used_bytes` bytes as offset for
where the blob would be stored. For most cases this was correct, however
if call to `reserve` (called after it) would need to `fallocate` more
space in the date file it could modify `used_bytes` if the last bytes of
an `mmap`-area should go used. This usually happens if the previous last
(before the call to `reserve`) area couldn't be expanded to accommodate
the new blob.

The solution is simple, rather then getting the offset before the call
to `reserve` to it after it.