Vote to fail consensus rather than abort in participant peer
Voting to abort means the operation can't proceed, e.g. when a blob is
already stored when trying to store it, or when trying to remove a blob
that is not stored.

Voting to fail means something went wrong, i.e. we've hit an unexpected
error, such as an disk or network error.