Change File::sync_all to File::sync_data in Index
On macOS the call is the same (fcntl(2) with F_FULLFSYNC), but on Linux
it uses fdatasync rather then fsync. Just calling fdatasync on Linux is
good enough because its doc say the following:

> fdatasync() is similar to fsync(), but does not flush modified
> metadata unless that metadata is needed in order to allow a subsequent
> data retrieval to be correctly handled.
[...]
> On the other hand, a change to the file size (st_size, as made by say
> ftruncate(2)), would require a metadata flush.

In other words it does all we need to do, ensure the next read (even
after a system crash) can read the index entry.