Redesign Protocol trait
Splits the Error type into two: RequestError and ResponseError.
Furthermore this adds the IsFatal trait requirement to RequestError to
determine whether or not an error is fatal. For example an I/O error
which can't be recovered from versus a protocol error which can be
recovered from.

Also add reply_to_error to send a reply for an error.