Add on-disk Storage implementation
This is an initial implementation, which will have bugs.

The storage is split into two files: the data file and the index file.
Both files are append-only logs. The index file determines what blobs
are actually in the storage, this means that even though the bytes might
be in the data file it doesn't mean that those bytes are in the storage.
The index file has the final say in what blobs are and aren't in the
storage.

See the documentation of the new storage::disk module for more
information.