Add new in-memory index
This new implementation is effectively a left-right hash map based on a
Hash Array Mapped Trie (HAMT). Some resources:
 * <https://en.wikipedia.org/wiki/Hash_array_mapped_trie>,
 * <https://idea.popcount.org/2012-07-25-introduction-to-hamt>,
 * Ideal Hash Trees by Phil Bagwell
 * Fast And Space Efficient Trie Searches by Phil Bagwell

The structure is implemented as a tree with `N_BRANCHES` (16) pointers on
each level. A pointer can point to a `Branch`, which again contains multiple
pointers, or point to [`Entry`].

Indexing into the structure is done using the [`Key`] of the blob, the bytes
are used as the index of the branch `LEVEL_SHIFT` (4) bits at the time.

This also gives us snapshots, which are a consistent view of the entry
index, without blocking the writer (in the context of the left-right
implementation).